project: use rust workspace and split in sub-projects
