+++ bitcoin/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "btctipserver-bitcoin"
+version = "0.1.1-dev"
+
+[dependencies]
+bdk = { version = "0.13.0", default-features = false }
+bdk-macros = "^0.6"
+log = "0.4.0"
+env_logger = "0.8.2"
+serde_json = { version = "^1.0" }
+dirs-next = "2.0.0"
+rust-ini = "0.16"
+structopt = "0.3"
+base64 = "0.13.0"
+http = "0.1.0"
+hex = "0.4"
+random-string = "1.0.0"
+
+[features]
+default = ["electrum", "bdk/key-value-db"]
+electrum = ["bdk/electrum"]
