+++ liquid/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "btctipserver-liquid"
+version = "0.1.1-dev"
+
+[dependencies]
+edk = { git = "https://github.com/lvaccaro/edk.git" }
+log = "0.4.0"
+env_logger = "0.8.2"
+serde_json = { version = "^1.0" }
+dirs-next = "2.0.0"
+rust-ini = "0.16"
+structopt = "0.3"
+base64 = "0.13.0"
+http = "0.1.0"
+hex = "0.4"
+random-string = "1.0.0"
+secp256k1-sys = "=0.4.1"
+
+[features]
+electrum = ["edk/electrum"]
